move MonitorResizeEvents and ReadRawStdin to terminal package
