+++ integration/apps_test.go
@@ -991,9 +991,7 @@ var _ = suite("apps/get-logs", func(t *testing.T, when spec.G, it spec.S) {
 				json.NewEncoder(buf).Encode(data)
 
 				err = c.WriteMessage(websocket.TextMessage, buf.Bytes())
-				if err != nil {
-					require.NoError(t, err)
-				}
+				require.NoError(t, err)
 
 				if i == finish {
 					break
