+++ pkg/listen/listen.go
@@ -6,11 +6,9 @@ import (
 	"fmt"
 	"io"
 	"net/url"
-	"os"
 
 	"github.com/gorilla/websocket"
 	"golang.org/x/sync/errgroup"
-	"golang.org/x/term"
 )
 
 // SchemaFunc takes a slice of bytes and returns an io.Reader (See Listener.SchemaFunc)
@@ -37,8 +35,6 @@ type Listener struct {
 // ListenerService listens to a websocket connection and outputs to the provided io.Writer
 type ListenerService interface {
 	Listen(ctx context.Context) error
-	ReadRawStdin(ctx context.Context, stdinCh chan<- string) error
-	MonitorResizeEvents(ctx context.Context, resizeEvents chan<- TerminalSize) error
 }
 
 var _ ListenerService = &Listener{}
@@ -116,28 +112,4 @@ func (l *Listener) Listen(ctx context.Context) error {
 		return err
 	}
 	return nil
-}
-
-// ReadRawStdin reads raw stdin.
-func (l *Listener) ReadRawStdin(ctx context.Context, stdinCh chan<- string) error {
-	// Set terminal to raw mode
-	oldState, err := term.MakeRaw(int(os.Stdin.Fd()))
-	if err != nil {
-		return fmt.Errorf("error setting terminal to raw mode: %v", err)
-	}
-	defer term.Restore(int(os.Stdin.Fd()), oldState) // Restore terminal on exit
-
-	for {
-		var b [1]byte
-		_, err := os.Stdin.Read(b[:]) // Read one byte at a time
-		if err != nil {
-			return fmt.Errorf("error reading stdin: %v", err)
-		}
-
-		select {
-		case stdinCh <- string(b[:]):
-		case <-ctx.Done():
-			return nil
-		}
-	}
 }
