+++ pkg/terminal/resize.go
@@ -1,33 +1,35 @@
-package listen
+//go:build !windows
+// +build !windows
+
+package terminal
 
 import (
 	"context"
 	"fmt"
 	"os"
-	"time"
+	"os/signal"
 
+	"golang.org/x/sys/unix"
 	"golang.org/x/term"
 )
 
 // MonitorResizeEvents monitors the terminal for resize events and sends them to the provided channel.
-func (l *Listener) MonitorResizeEvents(ctx context.Context, resizeEvents chan<- TerminalSize) error {
-	var prevTerminalSize TerminalSize
+func MonitorResizeEvents(ctx context.Context, resizeEvents chan<- TerminalSize) error {
+	winch := make(chan os.Signal, 1)
+	signal.Notify(winch, unix.SIGWINCH)
+	defer signal.Stop(winch)
 
-	ticker := time.NewTicker(250 * time.Millisecond)
 	for {
 		width, height, err := term.GetSize(int(os.Stdin.Fd()))
 		if err != nil {
 			return fmt.Errorf("error getting terminal size: %w", err)
 		}
 		terminalSize := TerminalSize{Width: width, Height: height}
-		if terminalSize != prevTerminalSize {
-			prevTerminalSize = terminalSize
-			resizeEvents <- terminalSize
-		}
 
-		// sleep to avoid hot looping
+		resizeEvents <- terminalSize
+
 		select {
-		case <-ticker.C:
+		case <-winch:
 		case <-ctx.Done():
 			return nil
 		}
