+++ pkg/terminal/terminalsize.go
@@ -1,4 +1,4 @@
-package listen
+package terminal
 
 // TerminalSize represents the size of a terminal
 type TerminalSize struct {
